Consolidated into extensions folder.
